# Backend/src/infrastructure/external/Docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Pagination": {
				"properties": {
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"total_pages": {
						"type": "number",
						"format": "double"
					},
					"hasNextPage": {
						"type": "boolean"
					},
					"hasPrevPage": {
						"type": "boolean"
					}
				},
				"required": [
					"total",
					"page",
					"limit",
					"total_pages",
					"hasNextPage",
					"hasPrevPage"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"create_permission_dto": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"description"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "Healthtech App",
		"description": "Healthtech App's Documentation",
		"version": "1.0.0",
		"contact": {
			"name": "Moises Armando Heredia",
			"email": "moisarm03@gmail.com"
		}
	},
	"paths": {
...
		"/permission": {
			"get": {
				"operationId": "Get_permissions",
				"responses": {
					"200": {
						"description": "Permissions retrieve succesfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"data": {
													"properties": {
														"pagination": {
															"$ref": "#/components/schemas/Pagination"
														},
														"data": {
															"items": {
																"properties": {
																	"description": {
																		"type": "string"
																	},
																	"id": {
																		"type": "number",
																		"format": "double"
																	},
																	"name": {
																		"type": "string"
																	}
																},
																"required": [
																	"description",
																	"id",
																	"name"
																],
																"type": "object"
															},
															"type": "array"
														}
													},
													"required": [
														"pagination",
														"data"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"data",
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {},
												"error": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/permission/{id}": {
			"get": {
				"operationId": "Search_permission",
				"responses": {
					"200": {
						"description": "Permission retrieve succesfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"data": {
													"properties": {
														"description": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														},
														"name": {
															"type": "string"
														}
													},
													"required": [
														"description",
														"id",
														"name"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"data",
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {},
												"error": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update_permission",
				"responses": {
					"200": {
						"description": "Permission updated succesfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"data": {
													"properties": {
														"description": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														},
														"name": {
															"type": "string"
														}
													},
													"required": [
														"description",
														"id",
														"name"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"data",
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {},
												"error": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/create_permission_dto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete_permission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/permission/new": {
			"post": {
				"operationId": "Create_permission",
				"responses": {
					"201": {
						"description": "Permission created succesfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"error": {},
												"data": {
													"properties": {
														"description": {
															"type": "string"
														},
														"id": {
															"type": "number",
															"format": "double"
														},
														"name": {
															"type": "string"
														}
													},
													"required": [
														"description",
														"id",
														"name"
													],
													"type": "object"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"data",
												"message",
												"status"
											],
											"type": "object"
										},
										{
											"properties": {
												"data": {},
												"error": {
													"type": "string"
												},
												"message": {
													"type": "string"
												},
												"status": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"error",
												"message",
												"status"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/create_permission_dto"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}